Polish sender and receiver closing
Fix logging in case of error in closing sequence, track open connection
and close it directly instead of relying on the connection mono.

References #103
