+++ src/main/java/reactor/rabbitmq/Utils.java
@@ -73,13 +73,6 @@ public abstract class Utils {
         R apply(T t) throws Exception;
     }
 
-    @FunctionalInterface
-    interface ExceptionRunnable {
-
-        void run() throws Exception;
-
-    }
-
     public static class SingleConnectionSupplier implements Utils.ExceptionFunction<ConnectionFactory, Connection> {
 
         private final Callable<? extends Connection> creationAction;
